Update key names; Separate method in Utility
