+++ Assets/AppCenter/Plugins/AppCenterSDK/Core/Android/AppCenterInternal.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 #if UNITY_ANDROID && !UNITY_EDITOR
+using Assets.AppCenter.Plugins.Android.Utility;
 using System;
 using System.Collections.Generic;
 using UnityEngine;
@@ -11,8 +12,6 @@ namespace Microsoft.AppCenter.Unity.Internal
     class AppCenterInternal
     {
         private static AndroidJavaClass _appCenter = new AndroidJavaClass("com.microsoft.appcenter.AppCenter");
-        private static AndroidJavaObject _context;
-        private static String PREFS_NAME = "AppCenterUserPrefs";
 
         public static void SetLogLevel(int logLevel)
         {
@@ -132,32 +131,12 @@ namespace Microsoft.AppCenter.Unity.Internal
             });
         }
 
-        private static AndroidJavaObject GetAndroidContext()
-        {
-            if (_context != null)
-            {
-                return _context;
-            }
-            var unityPlayer = new AndroidJavaClass("com.unity3d.player.UnityPlayer");
-            var activity = unityPlayer.GetStatic<AndroidJavaObject>("currentActivity");
-            _context = activity.Call<AndroidJavaObject>("getApplicationContext");
-            return _context;
-        }
-
-        public static void SetPreferenceInt(string prefKey, int prefValue) {
-            AndroidJavaObject context = GetAndroidContext();
-            AndroidJavaObject sharedPreferences = context.Call<AndroidJavaObject>("getSharedPreferences", new object[] { PREFS_NAME, 0 });
-            AndroidJavaObject editor = sharedPreferences.Call<AndroidJavaObject>("edit");
-            editor = editor.Call<AndroidJavaObject>("putInt", new object[] { prefKey, prefValue });
-            editor.Call("apply");
-        }
-
         public static void StartFromLibrary(IntPtr servicesArray)
         {
             var startMethod = AndroidJNI.GetStaticMethodID(_appCenter.GetRawClass(), "startFromLibrary", "(Landroid/content/Context;[Ljava/lang/Class;)V");
             AndroidJNI.CallStaticVoidMethod(_appCenter.GetRawClass(), startMethod, new jvalue[]
             {
-                new jvalue { l = GetAndroidContext().GetRawObject() },
+                new jvalue { l = AndroidUtility.GetAndroidContext().GetRawObject() },
                 new jvalue { l = servicesArray }
             });
         }
