+++ Assets/AppCenter/Plugins/AppCenterSDK/Core/Blank/AppCenterInternal.cs
@@ -37,10 +37,6 @@ namespace Microsoft.AppCenter.Unity.Internal
         {
         }
 
-        public static void SetPreferenceInt(string prefKey, int prefValue)
-        {
-        }
-
         public static void Start(string appSecret, ServiceType[] services)
         {
         }
