+++ Assets/AppCenter/Plugins/AppCenterSDK/Core/UWP/AppCenterInternal.cs
@@ -31,10 +31,6 @@ namespace Microsoft.AppCenter.Unity.Internal
             UWPAppCenter.Start(appSecret, nativeServiceTypes);
         }
 
-        public static void SetPreferenceInt(string prefKey, int prefValue) 
-        {
-        }
-
         public static void Start(Type[] services)
         {
             var nativeServiceTypes = ServicesToNativeTypes(services);
