+++ Assets/AppCenter/Plugins/iOS/Core/AppCenterStarter.original
@@ -40,7 +40,7 @@ static NSString *const kMSTargetToken = @"appcenter-transmission-target-token";
 static NSString *const kMSCustomLogUrl = @"custom-log-url";
 static NSString *const kMSCustomApiUrl = @"custom-api-url";
 static NSString *const kMSCustomInstallUrl = @"custom-install-url";
-static NSString *const kMSStartTargetKey = @"start-target-user-defaults-key";
+static NSString *const kMSStartTargetKey = @"MSAppCenterStartTargetUnityKey";
 
 static const int kMSLogLevel = 0/*LOG_LEVEL*/;
 static const int kMSStartupType = 0/*STARTUP_TYPE*/;
